comm: remove AddPemToCertPool
Use AppendCertsFromPEM method on CertPool in its place.
