+++ cmd/osnadmin/main.go
@@ -20,7 +20,6 @@ import (
 	"github.com/golang/protobuf/proto"
 	"github.com/hyperledger/fabric-protos-go/common"
 	"github.com/hyperledger/fabric/internal/osnadmin"
-	"github.com/hyperledger/fabric/internal/pkg/comm"
 	"github.com/hyperledger/fabric/protoutil"
 	"gopkg.in/alecthomas/kingpin.v2"
 )
@@ -81,9 +80,8 @@ func executeForArgs(args []string) (output string, exit int, err error) {
 		if err != nil {
 			return "", 1, fmt.Errorf("reading orderer CA certificate: %s", err)
 		}
-		err = comm.AddPemToCertPool(caFilePEM, caCertPool)
-		if err != nil {
-			return "", 1, fmt.Errorf("adding ca-file PEM to cert pool: %s", err)
+		if !caCertPool.AppendCertsFromPEM(caFilePEM) {
+			return "", 1, fmt.Errorf("failed to add ca-file PEM to cert pool")
 		}
 
 		tlsClientCert, err = tls.LoadX509KeyPair(*clientCert, *clientKey)
