+++ core/chaincode/extcc/extcc_handler_test.go
@@ -10,6 +10,7 @@ import (
 	"net"
 	"time"
 
+	"github.com/hyperledger/fabric/common/crypto/tlsgen"
 	"github.com/hyperledger/fabric/core/chaincode/extcc"
 	"github.com/hyperledger/fabric/core/chaincode/extcc/mock"
 	"github.com/hyperledger/fabric/core/container/ccintf"
@@ -75,15 +76,19 @@ var _ = Describe("Extcc", func() {
 		Context("chaincode info incorrect", func() {
 			var ccinfo *ccintf.ChaincodeServerInfo
 			BeforeEach(func() {
+				ca, err := tlsgen.NewCA()
+				Expect(err).NotTo(HaveOccurred())
+				ckp, err := ca.NewClientCertKeyPair()
+				Expect(err).NotTo(HaveOccurred())
 				ccinfo = &ccintf.ChaincodeServerInfo{
 					Address: "ccaddress:12345",
 					ClientConfig: comm.ClientConfig{
 						SecOpts: comm.SecureOptions{
 							UseTLS:            true,
 							RequireClientCert: true,
-							Certificate:       []byte("fake-cert"),
-							Key:               []byte("fake-key"),
-							ServerRootCAs:     [][]byte{[]byte("fake-root-cert")},
+							Certificate:       ckp.Cert,
+							Key:               ckp.Key,
+							ServerRootCAs:     [][]byte{ca.CertBytes()},
 						},
 						DialTimeout: 10 * time.Second,
 					},
