+++ internal/pkg/comm/creds.go
@@ -94,8 +94,7 @@ func (t *TLSConfig) SetClientCAs(certPool *x509.CertPool) {
 }
 
 // ClientHandShake is not implemented for `serverCreds`.
-func (sc *serverCreds) ClientHandshake(context.Context,
-	string, net.Conn) (net.Conn, credentials.AuthInfo, error) {
+func (sc *serverCreds) ClientHandshake(context.Context, string, net.Conn) (net.Conn, credentials.AuthInfo, error) {
 	return nil, nil, ErrClientHandshakeNotImplemented
 }
