+++ internal/pkg/comm/server_test.go
@@ -499,7 +499,7 @@ func TestNewGRPCServerInvalidParameters(t *testing.T) {
 
 	badRootCAs := [][]byte{[]byte(badPEM)}
 	err = srv.SetClientRootCAs(badRootCAs)
-	require.EqualError(t, err, "failed to set client root certificate(s): asn1: syntax error: data truncated")
+	require.EqualError(t, err, "failed to set client root certificate(s)")
 }
 
 func TestNewGRPCServer(t *testing.T) {
