+++ internal/pkg/comm/util.go
@@ -11,7 +11,6 @@ import (
 	"context"
 	"crypto/sha256"
 	"crypto/x509"
-	"encoding/pem"
 	"net"
 
 	"github.com/golang/protobuf/proto"
@@ -19,41 +18,6 @@ import (
 	"google.golang.org/grpc/credentials"
 	"google.golang.org/grpc/peer"
 )
-
-// AddPemToCertPool adds PEM-encoded certs to a cert pool
-func AddPemToCertPool(pemCerts []byte, pool *x509.CertPool) error {
-	certs, err := pemToX509Certs(pemCerts)
-	if err != nil {
-		return err
-	}
-	for _, cert := range certs {
-		pool.AddCert(cert)
-	}
-	return nil
-}
-
-// parse PEM-encoded certs
-func pemToX509Certs(pemCerts []byte) ([]*x509.Certificate, error) {
-	var certs []*x509.Certificate
-
-	// it's possible that multiple certs are encoded
-	for len(pemCerts) > 0 {
-		var block *pem.Block
-		block, pemCerts = pem.Decode(pemCerts)
-		if block == nil {
-			break
-		}
-
-		cert, err := x509.ParseCertificate(block.Bytes)
-		if err != nil {
-			return nil, err
-		}
-
-		certs = append(certs, cert)
-	}
-
-	return certs, nil
-}
 
 // BindingInspector receives as parameters a gRPC context and an Envelope,
 // and verifies whether the message contains an appropriate binding to the context
